README 20170211

+++ instapush_signal.py
@@ -1,17 +1,12 @@
-import os  
-import signal
+import os, signal
 from time import sleep
 pid = input("Input PID. ")
-bint = bin(input("Input a number. ")).replace("0b","")
-lbn = len(bint)
+bint = bin(input("Input a number. ")).replace("0b",""); lbn = len(bint)
 os.kill(pid,signal.SIGCONT)
-print "Binary:", bint
-sleep(0.7)
+print "Binary:", bint; sleep(0.7)
 for l in range (0, lbn):
-	if bint[l] == "0":
-		os.kill(pid,signal.SIGUSR1)
-	if bint[l] == "1":
-		os.kill(pid,signal.SIGUSR2)
+	if bint[l] == "0": os.kill(pid,signal.SIGUSR1)
+	if bint[l] == "1": os.kill(pid,signal.SIGUSR2)
 	print "Sending letter No." + str(l+1), ", " + str(lbn) + " in total."
 	sleep(0.3)
 os.kill(pid,signal.SIGTERM)
